requirements.txt
PySide6==6.8.2.1
PyYAML==6.0.1 